Move bitbake_variables fixture to common.py.
